Add the user interface
This is a very big commit again. I didn't continue to commit often after
starting work on the interface, as I was unsure of the steps I was
taking. I thought I should only commit once I am sure that what I am
committing will be part of the final output. This is not an excuse,
merely an explanation to my future self.

Commit every time you complete a piece of task.